first pass for classes
